GEODE-9065 - Introduce ClasspathService.
