+++ boms/geode-all-bom/src/test/resources/expected-pom.xml
@@ -903,12 +903,6 @@
       </dependency>
       <dependency>
         <groupId>org.apache.geode</groupId>
-        <artifactId>geode-deployment-legacy</artifactId>
-        <version>${version}</version>
-        <scope>compile</scope>
-      </dependency>
-      <dependency>
-        <groupId>org.apache.geode</groupId>
         <artifactId>geode-dunit</artifactId>
         <version>${version}</version>
         <scope>compile</scope>
@@ -1072,6 +1066,12 @@
       <dependency>
         <groupId>org.apache.geode</groupId>
         <artifactId>geode-modules-tomcat9</artifactId>
+        <version>${version}</version>
+        <scope>compile</scope>
+      </dependency>
+      <dependency>
+        <groupId>org.apache.geode</groupId>
+        <artifactId>geode-deployment-legacy</artifactId>
         <version>${version}</version>
         <scope>compile</scope>
       </dependency>
