+++ geode-assembly/src/distributedTest/java/org/apache/geode/management/internal/rest/DeploymentSemanticVersionJarDUnitTest.java
@@ -30,9 +30,9 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
-import org.apache.geode.deployment.internal.JarDeploymentServiceFactory;
 import org.apache.geode.distributed.internal.InternalConfigurationPersistenceService;
 import org.apache.geode.internal.classloader.ClassPathLoader;
+import org.apache.geode.internal.deployment.DeploymentServiceFactory;
 import org.apache.geode.management.api.ClusterManagementService;
 import org.apache.geode.management.api.RealizationResult;
 import org.apache.geode.management.client.ClusterManagementServiceBuilder;
@@ -218,7 +218,7 @@ public class DeploymentSemanticVersionJarDUnitTest {
 
   private static void verifyLoadAndHasVersion(String artifactId, String className, String version)
       throws Exception {
-    assertThat(JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    assertThat(DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .getDeployed(artifactId).isSuccessful()).isTrue();
     Class<?> klass = ClassPathLoader.getLatest().forName(className);
     assertThat(klass).isNotNull();
