+++ geode-core/build.gradle
@@ -320,6 +320,10 @@ dependencies {
     implementation(project(':geode-serialization'))
     implementation(project(':geode-tcp-server'))
 
+    runtimeOnly(project(':geode-deployment:geode-deployment-legacy')) {
+        ext.optional = true
+    }
+
     //geode-management currently has pieces of the public API
     //copied into it, so it is an API dependency
     api(project(':geode-management'))
