+++ geode-core/src/integrationTest/java/org/apache/geode/internal/classloader/ClassPathLoaderJUnitTest.java
@@ -26,7 +26,7 @@ import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import org.junit.rules.TestName;
 
-import org.apache.geode.deployment.internal.JarDeploymentServiceFactory;
+import org.apache.geode.internal.deployment.DeploymentServiceFactory;
 import org.apache.geode.management.configuration.Deployment;
 import org.apache.geode.test.compiler.ClassBuilder;
 
@@ -57,22 +57,22 @@ public class ClassPathLoaderJUnitTest {
     File jarA = createVersionOfJar(testName + "JarAVersion11", classAName, testName + "JarA.jar");
     File jarB = createVersionOfJar(testName + "JarBVersion11", classBName, testName + "JarB.jar");
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarA));
     Class class1 =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classAName);
     Object classAInstance = class1.newInstance();
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarB));
     Class class2 =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classAName);
     Object anotherClassAInstance = class2.newInstance();
 
     assertThat(classAInstance).isInstanceOf(anotherClassAInstance.getClass());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarA.getName());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarB.getName());
   }
 
@@ -97,9 +97,9 @@ public class ClassPathLoaderJUnitTest {
     File jarA = createJarFromClassContents(testName + "JarAVersion1", classAName,
         testName + "JarA.jar", classAContents,
         jarB.getAbsolutePath());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarA));
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarB));
 
     Class classA =
@@ -107,9 +107,9 @@ public class ClassPathLoaderJUnitTest {
     Object classAInstance = classA.newInstance();
 
     assertThat(classAName).isEqualTo(classAInstance.getClass().getSimpleName());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarA.getName());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarB.getName());
   }
 
@@ -136,9 +136,9 @@ public class ClassPathLoaderJUnitTest {
         testName + "JarA.jar", classAContents,
         jarB.getAbsolutePath());
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarA));
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarB));
 
     Class classA =
@@ -150,9 +150,9 @@ public class ClassPathLoaderJUnitTest {
     Object classBInstance = classB.newInstance();
 
     assertThat(classBInstance.getClass()).isAssignableFrom(classAInstance.getClass());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarA.getName());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarB.getName());
   }
 
@@ -180,13 +180,13 @@ public class ClassPathLoaderJUnitTest {
         testName + "JarA.jar", classAContents,
         jarB.getAbsolutePath());
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarB));
     Class classB =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classBName);
     Object classBInstance = classB.newInstance();
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarA));
     Class classA =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classAName);
@@ -198,9 +198,9 @@ public class ClassPathLoaderJUnitTest {
 
     assertThat(classB2Instance).isInstanceOf(classBInstance.getClass());
     assertThat(classBInstance.getClass()).isAssignableFrom(classAInstance.getClass());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarB.getName());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarA.getName());
   }
 
@@ -239,29 +239,29 @@ public class ClassPathLoaderJUnitTest {
             classA2Contents,
             jarB.getAbsolutePath());
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarB));
     Class classB =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classBName);
 
     assertThat(
-        JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+        DeploymentServiceFactory.getJarDeploymentServiceInstance()
             .deploy(createDeploymentFromJar(jarA)))
                 .isNotNull();
     Class classA =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classAName);
     Object classAInstance = classA.newInstance();
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarAV2));
     Class classA2 =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classAName);
     Object classA2Instance = classA2.newInstance();
 
     assertThat(classA2Instance).isNotInstanceOf(classA);
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarB.getName());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarA.getName());
   }
 
@@ -297,20 +297,20 @@ public class ClassPathLoaderJUnitTest {
             classB2Contents);
 
     assertThat(
-        JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+        DeploymentServiceFactory.getJarDeploymentServiceInstance()
             .deploy(createDeploymentFromJar(jarB)))
                 .isNotNull();
     Class classB =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classBName);
     Object classBInstance = classB.newInstance();
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarA));
     Class classA =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classAName);
     Object classAInstance = classA.newInstance();
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarBV2));
 
     Class classA2 =
@@ -318,9 +318,9 @@ public class ClassPathLoaderJUnitTest {
     Object classA2Instance = classA2.newInstance();
 
     assertThat(classA2Instance).isInstanceOf(classAInstance.getClass());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarA.getName());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarBV2.getName());
   }
 
@@ -356,10 +356,10 @@ public class ClassPathLoaderJUnitTest {
         createJarFromClassContents(testName + "JarBVersion2", classBName, testName + "JarB.jar",
             classB2Contents);
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarA));
     assertThat(
-        JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+        DeploymentServiceFactory.getJarDeploymentServiceInstance()
             .deploy(createDeploymentFromJar(jarB)))
                 .isNotNull();
 
@@ -371,16 +371,16 @@ public class ClassPathLoaderJUnitTest {
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classBName);
     Object classBInstance = classB.newInstance();
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarBV2));
     Class classA2 =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classAName);
     Object classA2Instance = classA2.newInstance();
 
     assertThat(classA2Instance).isInstanceOf(classAInstance.getClass());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarA.getName());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarB.getName());
   }
 
@@ -417,13 +417,13 @@ public class ClassPathLoaderJUnitTest {
         createJarFromClassContents(testName + "JarBVersion2", classB2Name, testName + "JarB.jar",
             classB2Contents);
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarB));
     Class classB =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classBName);
     Object classBInstance = classB.newInstance();
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarBV2));
 
     try {
@@ -432,7 +432,7 @@ public class ClassPathLoaderJUnitTest {
     } catch (ClassNotFoundException ex) {
       // Expected
     } finally {
-      JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+      DeploymentServiceFactory.getJarDeploymentServiceInstance()
           .undeployByFileName(jarB.getName());
     }
   }
@@ -459,24 +459,24 @@ public class ClassPathLoaderJUnitTest {
         testName + "JarA.jar", classAContents);
 
     assertThat(
-        JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+        DeploymentServiceFactory.getJarDeploymentServiceInstance()
             .deploy(createDeploymentFromJar(jarB)))
                 .isNotNull();
     Class classB =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classBName);
     Object classBInstance = classB.newInstance();
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarA));
     Class classA =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classAName);
     Object classAInstance = classA.newInstance();
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(testName + "JarB.jar");
     Object classA2Instance = classA.newInstance();
     assertThat(classA2Instance).isInstanceOf(classAInstance.getClass());
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(jarA.getName());
   }
 
@@ -509,21 +509,21 @@ public class ClassPathLoaderJUnitTest {
         jarB.getAbsolutePath());
 
     assertThat(
-        JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+        DeploymentServiceFactory.getJarDeploymentServiceInstance()
             .deploy(createDeploymentFromJar(jarB)))
                 .isNotNull();
     Class classB =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classBName);
     Object classBInstance = classB.newInstance();
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .deploy(createDeploymentFromJar(jarA));
 
     Class classA =
         ClassPathLoader.getLatest().asClassLoader().loadClass("jddunit.function." + classAName);
     Object classAInstance = classA.newInstance();
 
-    JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .undeployByFileName(testName + "JarB.jar");
     try {
       classB =
@@ -532,7 +532,7 @@ public class ClassPathLoaderJUnitTest {
     } catch (ClassNotFoundException ex) {
       // expected
     } finally {
-      JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+      DeploymentServiceFactory.getJarDeploymentServiceInstance()
           .undeployByFileName(jarA.getName());
     }
