+++ geode-core/src/main/java/org/apache/geode/cache/internal/execute/FunctionToFileTracker.java
@@ -16,9 +16,7 @@
  */
 package org.apache.geode.cache.internal.execute;
 
-import java.io.BufferedInputStream;
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Modifier;
@@ -29,8 +27,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.jar.JarEntry;
-import java.util.jar.JarInputStream;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
@@ -66,51 +62,29 @@ public class FunctionToFileTracker {
    * @param jarFile The {@link File} that contains {@link Function} that are to be registered
    */
   public synchronized void registerFunctionsFromFile(String deploymentName, File jarFile)
-      throws ClassNotFoundException {
+      throws ClassNotFoundException, IOException {
     logger.debug("Registering functions for: {}", jarFile.getName());
 
     List<Function<?>> registeredFunctions = new LinkedList<>();
-
-    try (BufferedInputStream bufferedInputStream =
-        new BufferedInputStream(new FileInputStream(jarFile))) {
-      try (JarInputStream jarInputStream = new JarInputStream(bufferedInputStream)) {
+    try {
+      Collection<String> functionClasses = findFunctionsInThisJar(jarFile);
+      String filePath = jarFile.getAbsolutePath();
+      Collection<Function<?>> functions = new LinkedList<>();
+      for (String functionClass : functionClasses) {
+        logger.debug("Attempting to load class: {}, from JAR file: {}", functionClass,
+            filePath);
         try {
-          Collection<String> functionClasses = findFunctionsInThisJar(jarFile);
-          JarEntry jarEntry = jarInputStream.getNextJarEntry();
-          String filePath = jarFile.getAbsolutePath();
-          while (jarEntry != null) {
-            if (jarEntry.getName().endsWith(".class")) {
-              String className = PATTERN_SLASH.matcher(jarEntry.getName()).replaceAll("\\.")
-                  .substring(0, jarEntry.getName().length() - 6);
-
-              if (functionClasses.contains(className)) {
-                logger.debug("Attempting to load class: {}, from JAR file: {}", jarEntry.getName(),
-                    filePath);
-                try {
-                  Collection<Function<?>> functions = loadFunctionFromClassName(className);
-                  registeredFunctions.addAll(registerFunction(filePath, functions));
-                } catch (ClassNotFoundException | NoClassDefFoundError cnfex) {
-                  logger.error("Unable to load all classes from JAR file: {}",
-                      filePath, cnfex);
-                  throw cnfex;
-                }
-              } else {
-                logger.debug("No functions found in class: {}, from JAR file: {}",
-                    jarEntry.getName(),
-                    filePath);
-              }
-            }
-            jarEntry = jarInputStream.getNextJarEntry();
-          }
-        } catch (IOException ioex) {
-          logger.error("Exception when trying to read class from ByteArrayInputStream", ioex);
+          functions.addAll(loadFunctionFromClassName(functionClass));
+        } catch (ClassNotFoundException | NoClassDefFoundError cnfex) {
+          logger.error("Unable to load all classes from JAR file: {}",
+              filePath, cnfex);
+          throw cnfex;
         }
-      } catch (IOException ioex) {
-        logger.error("Exception attempting to close JAR input stream", ioex);
       }
-    } catch (Exception ex) {
-      logger.error("Unable to scan jar file for functions");
-      return;
+      registeredFunctions.addAll(registerFunction(filePath, functions));
+    } catch (IOException ioex) {
+      logger.error("Exception when trying to find function classes from Jar", ioex);
+      throw ioex;
     }
     List<Function<?>> previouslyRegisteredFunctions =
         deploymentToFunctionsMap.remove(deploymentName);
