+++ geode-core/src/main/java/org/apache/geode/internal/cache/backup/BackupFileCopier.java
@@ -30,12 +30,12 @@ import org.apache.commons.io.FileUtils;
 import org.apache.logging.log4j.Logger;
 
 import org.apache.geode.cache.DiskStore;
-import org.apache.geode.deployment.internal.JarDeploymentService;
 import org.apache.geode.distributed.DistributedSystem;
 import org.apache.geode.internal.cache.DirectoryHolder;
 import org.apache.geode.internal.cache.DiskStoreImpl;
 import org.apache.geode.internal.cache.InternalCache;
 import org.apache.geode.internal.cache.Oplog;
+import org.apache.geode.internal.deployment.JarDeploymentService;
 import org.apache.geode.internal.lang.SystemUtils;
 import org.apache.geode.logging.internal.log4j.api.LogService;
 import org.apache.geode.management.configuration.Deployment;
