+++ geode-core/src/main/java/org/apache/geode/internal/cache/backup/BackupTask.java
@@ -26,10 +26,10 @@ import org.apache.logging.log4j.Logger;
 import org.apache.geode.InternalGemFireError;
 import org.apache.geode.cache.DiskStore;
 import org.apache.geode.cache.persistence.PersistentID;
-import org.apache.geode.deployment.internal.JarDeploymentServiceFactory;
 import org.apache.geode.internal.cache.DiskStoreImpl;
 import org.apache.geode.internal.cache.InternalCache;
 import org.apache.geode.internal.cache.Oplog;
+import org.apache.geode.internal.deployment.DeploymentServiceFactory;
 import org.apache.geode.logging.internal.log4j.api.LogService;
 
 /**
@@ -108,7 +108,7 @@ class BackupTask {
       Collection<DiskStore> diskStores = cache.listDiskStoresIncludingRegionOwned();
       temporaryFiles = TemporaryBackupFiles.create();
       fileCopier = new BackupFileCopier(cache,
-          JarDeploymentServiceFactory.getJarDeploymentServiceInstance(), temporaryFiles);
+          DeploymentServiceFactory.getJarDeploymentServiceInstance(), temporaryFiles);
 
       Map<DiskStoreImpl, DiskStoreBackup> backupByDiskStores = startDiskStoreBackups(diskStores);
