+++ geode-core/src/main/java/org/apache/geode/internal/classloader/ClassPathLoader.java
@@ -14,30 +14,16 @@
  */
 package org.apache.geode.internal.classloader;
 
-import static java.util.stream.Collectors.joining;
-
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.lang.reflect.Proxy;
-import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.LinkedHashSet;
-import java.util.List;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 import org.apache.geode.annotations.VisibleForTesting;
 import org.apache.geode.annotations.internal.MakeNotStatic;
-import org.apache.geode.deployment.internal.JarDeploymentService;
-import org.apache.geode.deployment.internal.JarDeploymentServiceFactory;
-import org.apache.geode.internal.util.CollectionUtils;
-import org.apache.geode.management.configuration.Deployment;
+import org.apache.geode.internal.deployment.DeploymentServiceFactory;
+import org.apache.geode.internal.deployment.JarDeploymentService;
 import org.apache.geode.util.internal.GeodeGlossary;
 
 /**
@@ -73,34 +59,27 @@ import org.apache.geode.util.internal.GeodeGlossary;
  * @since GemFire 6.5.1.4
  */
 public class ClassPathLoader {
-  private static final Logger logger = LogManager.getLogger();
-
   static final String EXCLUDE_TCCL_PROPERTY =
       GeodeGlossary.GEMFIRE_PREFIX + "excludeThreadContextClassLoader";
-
   @MakeNotStatic
   private static volatile ClassPathLoader latest;
 
-  private final HashMap<String, DeployJarChildFirstClassLoader> artifactIdsToClassLoader =
-      new HashMap<>();
-
-  private volatile DeployJarChildFirstClassLoader leafLoader;
-
   private final JarDeploymentService jarDeploymentService;
+  private final ClasspathService classPathService;
 
-  private final boolean excludeTCCL;
 
   private ClassPathLoader(boolean excludeTCCL) {
-    this.excludeTCCL = excludeTCCL;
-    jarDeploymentService = JarDeploymentServiceFactory.getJarDeploymentServiceInstance();
-    rebuildClassLoaderForDeployedJars();
+    jarDeploymentService = DeploymentServiceFactory.getJarDeploymentServiceInstance();
+    classPathService = DeploymentServiceFactory.getClasspathServiceInstance();
+    classPathService.init(excludeTCCL, jarDeploymentService);
+
   }
 
   private ClassPathLoader(boolean excludeTCCL, File workingDir) {
-    this.excludeTCCL = excludeTCCL;
-    jarDeploymentService = JarDeploymentServiceFactory
+    jarDeploymentService = DeploymentServiceFactory
         .reinitializeJarDeploymentServiceWithWorkingDirectory(workingDir);
-    rebuildClassLoaderForDeployedJars();
+    classPathService = DeploymentServiceFactory.getClasspathServiceInstance();
+    classPathService.init(excludeTCCL, jarDeploymentService);
   }
 
   @VisibleForTesting
@@ -122,239 +101,31 @@ public class ClassPathLoader {
     return new ClassPathLoader(excludeTCCL);
   }
 
-  private synchronized void rebuildClassLoaderForDeployedJars() {
-    leafLoader = null;
-    List<Deployment> deployments = jarDeploymentService.listDeployed();
-    for (Deployment deployment : deployments) {
-      chainClassloader(deployment.getFile(), deployment.getDeploymentName());
-    }
-  }
-
-  private ClassLoader getLeafLoader() {
-    if (leafLoader == null) {
-      return ClassPathLoader.class.getClassLoader();
-    }
-    return leafLoader;
-  }
-
-  public synchronized void chainClassloader(File jar, String deploymentName) {
-    try {
-      leafLoader = new DeployJarChildFirstClassLoader(artifactIdsToClassLoader,
-          new URL[] {jar.toURI().toURL()}, deploymentName, getLeafLoader());
-    } catch (MalformedURLException e) {
-      e.printStackTrace();
-    }
-  }
-
-  public synchronized void unloadClassloaderForArtifact(String artifactId) {
-    artifactIdsToClassLoader.put(artifactId, null);
-  }
-
-  public URL getResource(final String name) {
-    final boolean isTraceEnabled = logger.isTraceEnabled();
-    if (isTraceEnabled) {
-      logger.trace("getResource({})", name);
-    }
-
-    for (ClassLoader classLoader : getClassLoaders()) {
-      if (isTraceEnabled) {
-        logger.trace("getResource trying: {}", classLoader);
-      }
-      try {
-        URL url = classLoader.getResource(name);
-
-        if (url != null) {
-          if (isTraceEnabled) {
-            logger.trace("getResource found by: {}", classLoader);
-          }
-          return url;
+  public static ClassPathLoader getLatest() {
+    if (latest == null) {
+      synchronized (ClassPathLoader.class) {
+        if (latest == null) {
+          setLatestToDefault();
         }
-      } catch (SecurityException e) {
-        // try next classLoader
       }
     }
 
-    return null;
-  }
-
-  public Class<?> forName(final String name) throws ClassNotFoundException {
-    final boolean isTraceEnabled = logger.isTraceEnabled();
-    if (isTraceEnabled) {
-      logger.trace("forName({})", name);
-    }
-
-    Class<?> clazz = forName(name, isTraceEnabled);
-    if (clazz != null)
-      return clazz;
-
-    throw new ClassNotFoundException(name);
-  }
-
-  private Class<?> forName(String name, boolean isTraceEnabled) {
-    for (ClassLoader classLoader : getClassLoaders()) {
-      if (isTraceEnabled) {
-        logger.trace("forName trying: {}", classLoader);
-      }
-      try {
-        // Do not look up class definitions in jars that have been unloaded or are old
-        if (classLoader instanceof DeployJarChildFirstClassLoader) {
-          if (((DeployJarChildFirstClassLoader) classLoader).thisIsOld()) {
-            return null;
-          }
-        }
-        Class<?> clazz = Class.forName(name, true, classLoader);
-        if (clazz != null) {
-          if (isTraceEnabled) {
-            logger.trace("forName found by: {}", classLoader);
-          }
-          return clazz;
-        }
-      } catch (SecurityException | ClassNotFoundException e) {
-        // try next classLoader
-      }
-    }
-    return null;
+    return latest;
   }
 
   /**
-   * See {@link Proxy#getProxyClass(ClassLoader, Class...)}
+   * Helper method equivalent to {@code ClassPathLoader.getLatest().asClassLoader();}.
+   *
+   * @return a ClassLoader for current ClassPathLoader.
+   * @since GemFire 8.1
    */
-  public Class<?> getProxyClass(final Class<?>... classObjs) {
-    IllegalArgumentException ex = null;
-
-    for (ClassLoader classLoader : getClassLoaders()) {
-      try {
-        return Proxy.getProxyClass(classLoader, classObjs);
-      } catch (SecurityException sex) {
-        // Continue to next classloader
-      } catch (IllegalArgumentException iaex) {
-        ex = iaex;
-        // Continue to next classloader
-      }
-    }
-
-    if (ex != null) {
-      throw ex;
-    }
-    return null;
+  public static ClassLoader getLatestAsClassLoader() {
+    return getLatest().asClassLoader();
   }
 
   @Override
   public String toString() {
-    final StringBuilder sb = new StringBuilder(getClass().getName());
-    sb.append("@").append(System.identityHashCode(this)).append("{");
-    sb.append("excludeTCCL=").append(excludeTCCL);
-    sb.append(", jarDeployer=").append(jarDeploymentService);
-    sb.append(", classLoaders=[");
-    sb.append(getClassLoaders().stream().map(ClassLoader::toString).collect(joining(", ")));
-    sb.append("]}");
-    return sb.toString();
-  }
-
-  /**
-   * Finds the resource with the given name. This method will first search the class loader of the
-   * context class for the resource. That failing, this method will invoke
-   * {@link #getResource(String)} to find the resource.
-   *
-   * @param contextClass The class whose class loader will first be searched
-   * @param name The resource name
-   * @return A <tt>URL</tt> object for reading the resource, or <tt>null</tt> if the resource could
-   *         not be found or the invoker doesn't have adequate privileges to get the resource.
-   */
-  public URL getResource(final Class<?> contextClass, final String name) {
-    if (contextClass != null) {
-      URL url = contextClass.getResource(name);
-      if (url != null) {
-        return url;
-      }
-    }
-    return getResource(name);
-  }
-
-  /**
-   * Returns an input stream for reading the specified resource.
-   * <p>
-   * The search order is described in the documentation for {@link #getResource(String)}.
-   *
-   * @param name The resource name
-   * @return An input stream for reading the resource, or <tt>null</tt> if the resource could not be
-   *         found
-   */
-  public InputStream getResourceAsStream(final String name) {
-    URL url = getResource(name);
-    try {
-      return url != null ? url.openStream() : null;
-    } catch (IOException e) {
-      return null;
-    }
-  }
-
-  /**
-   * Returns an input stream for reading the specified resource.
-   * <p>
-   * The search order is described in the documentation for {@link #getResource(Class, String)}.
-   *
-   * @param contextClass The class whose class loader will first be searched
-   * @param name The resource name
-   * @return An input stream for reading the resource, or <tt>null</tt> if the resource could not be
-   *         found
-   */
-  public InputStream getResourceAsStream(final Class<?> contextClass, final String name) {
-    if (contextClass != null) {
-      InputStream is = contextClass.getResourceAsStream(name);
-      if (is != null) {
-        return is;
-      }
-    }
-    return getResourceAsStream(name);
-  }
-
-  /**
-   * Finds all the resources with the given name. This method will first search the class loader of
-   * the context class for the resource before searching all other {@code ClassLoader}s.
-   *
-   * @param contextClass The class whose class loader will first be searched
-   * @param name The resource name
-   * @return An enumeration of <tt>URL</tt> objects for the resource. If no resources could be
-   *         found, the enumeration will be empty. Resources that the class loader doesn't have
-   *         access to will not be in the enumeration.
-   * @throws IOException If I/O errors occur
-   * @see ClassLoader#getResources(String)
-   */
-  private Enumeration<URL> getResources(final Class<?> contextClass, final String name)
-      throws IOException {
-    final LinkedHashSet<URL> urls = new LinkedHashSet<>();
-
-    if (contextClass != null) {
-      CollectionUtils.addAll(urls, contextClass.getClassLoader().getResources(name));
-    }
-
-    for (ClassLoader classLoader : getClassLoaders()) {
-      Enumeration<URL> resources = classLoader.getResources(name);
-      if (resources != null && resources.hasMoreElements()) {
-        CollectionUtils.addAll(urls, resources);
-      }
-    }
-
-    return Collections.enumeration(urls);
-  }
-
-  public Enumeration<URL> getResources(final String name) throws IOException {
-    return getResources(null, name);
-  }
-
-  private List<ClassLoader> getClassLoaders() {
-    ArrayList<ClassLoader> classLoaders = new ArrayList<>();
-
-    if (!excludeTCCL) {
-      ClassLoader tccl = Thread.currentThread().getContextClassLoader();
-      if (tccl != null) {
-        classLoaders.add(tccl);
-      }
-    }
-
-    classLoaders.add(getLeafLoader());
-    return classLoaders;
+    return classPathService.toString();
   }
 
   /**
@@ -392,25 +163,39 @@ public class ClassPathLoader {
     };
   }
 
-  public static ClassPathLoader getLatest() {
-    if (latest == null) {
-      synchronized (ClassPathLoader.class) {
-        if (latest == null)
-          setLatestToDefault();
-      }
-    }
-
-    return latest;
+  public void chainClassloader(File jar, String deploymentName) {
+    classPathService.chainClassloader(jar, deploymentName);
   }
 
-  /**
-   * Helper method equivalent to {@code ClassPathLoader.getLatest().asClassLoader();}.
-   *
-   * @return a ClassLoader for current ClassPathLoader.
-   * @since GemFire 8.1
-   */
-  public static ClassLoader getLatestAsClassLoader() {
-    return getLatest().asClassLoader();
+  public void unloadClassloaderForArtifact(String artifactId) {
+    classPathService.unloadClassloaderForArtifact(artifactId);
   }
 
+  public Class<?> forName(String name) throws ClassNotFoundException {
+    return classPathService.forName(name);
+  }
+
+  public Class<?> getProxyClass(Class<?>... classObjs) {
+    return classPathService.getProxyClass(classObjs);
+  }
+
+  public URL getResource(Class<?> contextClass, String name) {
+    return classPathService.getResource(contextClass, name);
+  }
+
+  public URL getResource(String name) {
+    return classPathService.getResource(name);
+  }
+
+  public InputStream getResourceAsStream(Class<?> contextClass, String name) {
+    return classPathService.getResourceAsStream(contextClass, name);
+  }
+
+  public InputStream getResourceAsStream(String name) {
+    return classPathService.getResourceAsStream(name);
+  }
+
+  public Enumeration<URL> getResources(String name) throws IOException {
+    return classPathService.getResources(name);
+  }
 }
