+++ geode-core/src/main/java/org/apache/geode/internal/deployment/DeploymentServiceFactory.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.geode.deployment.internal;
+package org.apache.geode.internal.deployment;
 
 import java.io.File;
 import java.util.ServiceLoader;
@@ -24,31 +24,48 @@ import org.apache.logging.log4j.Logger;
 import org.apache.geode.annotations.Experimental;
 import org.apache.geode.annotations.Immutable;
 import org.apache.geode.internal.cache.client.protocol.exception.ServiceLoadingFailureException;
+import org.apache.geode.internal.classloader.ClasspathService;
 import org.apache.geode.logging.internal.log4j.api.LogService;
 
 /**
- * Factory responsible for loading and holding the singleton instance of the
- * {@link JarDeploymentService}.
- * Use this class to retrieve the instance rather than creating one.
+ * Factory responsible for loading and holding the singleton instances of the
+ * {@link JarDeploymentService} and {@link ClasspathService}.
+ * Use this class to retrieve the instances rather than creating one.
  *
  * @since Geode 1.15
  */
 @Experimental
-public class JarDeploymentServiceFactory {
+public class DeploymentServiceFactory {
 
   private static final Logger logger = LogService.getLogger();
 
   @Immutable
   private static final JarDeploymentService jarDeploymentService = createJarDeploymentService();
 
+  @Immutable
+  private static final ClasspathService classpathService = createClassPathService();
+
   private static JarDeploymentService createJarDeploymentService() {
     ServiceLoader<JarDeploymentService> jarDeploymentServices =
         ServiceLoader.load(JarDeploymentService.class);
     if (jarDeploymentServices.iterator().hasNext()) {
-      return jarDeploymentServices.iterator().next();
+      JarDeploymentService jarDeploymentService = jarDeploymentServices.iterator().next();
+      return jarDeploymentService;
     } else {
       throw new ServiceLoadingFailureException(
           "No implementation of JarDeploymentService could be loaded.");
+    }
+  }
+
+  private static ClasspathService createClassPathService() {
+    ServiceLoader<ClasspathService> jarDeploymentServices =
+        ServiceLoader.load(ClasspathService.class);
+    if (jarDeploymentServices.iterator().hasNext()) {
+      ClasspathService classpathService = jarDeploymentServices.iterator().next();
+      return classpathService;
+    } else {
+      throw new ServiceLoadingFailureException(
+          "No implementation of ClasspathService could be loaded.");
     }
   }
 
@@ -59,6 +76,15 @@ public class JarDeploymentServiceFactory {
    */
   public static JarDeploymentService getJarDeploymentServiceInstance() {
     return jarDeploymentService;
+  }
+
+  /**
+   * Gets the current instance of the {@link ClasspathService}.
+   *
+   * @return current instance of the {@link ClasspathService}.
+   */
+  public static ClasspathService getClasspathServiceInstance() {
+    return classpathService;
   }
 
   /**
