+++ geode-core/src/test/java/org/apache/geode/internal/classloader/ClassPathLoaderTest.java
@@ -41,8 +41,8 @@ import org.junit.Test;
 import org.junit.contrib.java.lang.system.RestoreSystemProperties;
 import org.junit.rules.TemporaryFolder;
 
-import org.apache.geode.deployment.internal.JarDeploymentService;
-import org.apache.geode.deployment.internal.JarDeploymentServiceFactory;
+import org.apache.geode.internal.deployment.DeploymentServiceFactory;
+import org.apache.geode.internal.deployment.JarDeploymentService;
 import org.apache.geode.management.configuration.Deployment;
 import org.apache.geode.test.compiler.ClassBuilder;
 
@@ -83,7 +83,7 @@ public class ClassPathLoaderTest {
     zeroFile.createNewFile();
 
     JarDeploymentService jarDeploymentService =
-        JarDeploymentServiceFactory.getJarDeploymentServiceInstance();
+        DeploymentServiceFactory.getJarDeploymentServiceInstance();
     assertThatThrownBy(() -> {
       Deployment deployment =
           new Deployment("JarDeployerDUnitZLF.jar", "test", Instant.now().toString());
