+++ geode-deployment/geode-deployment-legacy/src/main/resources/META-INF/services/org.apache.geode.internal.classloader.ClasspathService
@@ -1,0 +1,1 @@
+org.apache.geode.classloader.internal.LegacyClasspathServiceImpl
