+++ geode-gfsh/src/distributedTest/java/org/apache/geode/management/internal/cli/commands/DeploySemanticVersionJarDUnitTest.java
@@ -28,9 +28,9 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
-import org.apache.geode.deployment.internal.JarDeploymentServiceFactory;
 import org.apache.geode.distributed.internal.InternalConfigurationPersistenceService;
 import org.apache.geode.internal.classloader.ClassPathLoader;
+import org.apache.geode.internal.deployment.DeploymentServiceFactory;
 import org.apache.geode.test.compiler.JarBuilder;
 import org.apache.geode.test.dunit.rules.ClusterStartupRule;
 import org.apache.geode.test.dunit.rules.MemberVM;
@@ -83,7 +83,7 @@ public class DeploySemanticVersionJarDUnitTest {
 
   private static void verifyLoadAndHasVersion(String artifactId, String className, String version)
       throws Exception {
-    assertThat(JarDeploymentServiceFactory.getJarDeploymentServiceInstance()
+    assertThat(DeploymentServiceFactory.getJarDeploymentServiceInstance()
         .getDeployed(artifactId).isSuccessful()).isTrue();
     Class<?> klass = ClassPathLoader.getLatest().forName(className);
     assertThat(klass).isNotNull();
