+++ geode-gfsh/src/distributedTest/java/org/apache/geode/management/internal/cli/commands/DeployWithGroupsDUnitTest.java
@@ -26,8 +26,8 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
-import org.apache.geode.deployment.internal.JarDeploymentServiceFactory;
 import org.apache.geode.internal.classloader.ClassPathLoader;
+import org.apache.geode.internal.deployment.DeploymentServiceFactory;
 import org.apache.geode.management.internal.utils.JarFileUtils;
 import org.apache.geode.test.compiler.ClassBuilder;
 import org.apache.geode.test.dunit.rules.ClusterStartupRule;
@@ -326,7 +326,7 @@ public class DeployWithGroupsDUnitTest implements Serializable {
   private void assertThatCanLoad(String jarName, String className) throws ClassNotFoundException {
     String deploymentName = JarFileUtils.getArtifactId(jarName);
     assertThat(
-        JarDeploymentServiceFactory.getJarDeploymentServiceInstance().getDeployed(deploymentName)
+        DeploymentServiceFactory.getJarDeploymentServiceInstance().getDeployed(deploymentName)
             .isSuccessful())
                 .isTrue();
     assertThat(ClassPathLoader.getLatest().forName(className)).isNotNull();
@@ -335,7 +335,7 @@ public class DeployWithGroupsDUnitTest implements Serializable {
   private void assertThatCannotLoad(String jarName, String className) {
     String deploymentName = JarFileUtils.getArtifactId(jarName);
     assertThat(
-        JarDeploymentServiceFactory.getJarDeploymentServiceInstance().getDeployed(deploymentName)
+        DeploymentServiceFactory.getJarDeploymentServiceInstance().getDeployed(deploymentName)
             .isSuccessful())
                 .isFalse();
     assertThatThrownBy(() -> ClassPathLoader.getLatest().forName(className))
