+++ gradle.properties
@@ -74,13 +74,12 @@ testJVMVer=8
 
 repeat = 100
 
-org.gradle.caching = true
+#org.gradle.caching = true
 org.gradle.configureondemand = false
-org.gradle.daemon = true
+#org.gradle.daemon = true
 org.gradle.jvmargs = -Xmx3g
-org.gradle.parallel = true
-#org.gradle.workers.max = 3
+org.gradle.parallel = trueJarDeployerDeadlockTestJarDeployerDeadlockTest
+org.gradle.workers.max = 3
 
 org.gradle.internal.http.socketTimeout=120000
 org.gradle.internal.http.connectionTimeout=120000
-DeploySemanticVersionJarDUnitTest
